+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Remove `Blob` type
+* Breaking changes to `TypeInfo` to properly support generics. Generics support was added in v0.2.2, but the implementation was incorrect.
+<!-- TODO: elaborate changes -->
+
 ## v0.2.3
 
 * Add changelog